Push functions to another file
