+++ d1.workflow.frontend.js
@@ -1,3 +1,5 @@
+import { makePostCall, addNameList, findPos, mapAssignee, mapFromElement} from "./d1.workflow.util.js"
+
 (function(jq$){
 	'use strict';
 	//jq$ is jQuery
@@ -23,8 +25,13 @@
         var names = mapAssignee(selected_office, selected_cu);
 
         //Creating Table view
-        jq$('[name="programOfficeId"]').closest('td').addClass("alignBar");
-        jq$('[name="programOfficeId"]').closest('tr').append('<td id="flexcontainer"> <div id="flexbox">'
+        //jq$('[name="programOfficeId"]').closest('td').addClass("alignBar");
+        jq$('[name="programOfficeId"]').closest('tr').append('<td id="flexcontainer">'
+        + '<div id="changeAssignment"> <div id="button-wrapper">'
+        + '<button type="button" id="assigneeList" class="yui3-button"> Assignee </button>'
+        + '<button type="button" id="wfOwnerList" class="yui3-button"> Workflow Owner </button> </div></div>'
+
+        + '<div id="flexbox">'
         + '<div class="user-flexchild"> </div>'
         + '<div class="assigned-flexchild"> </div>'
         + '</div></td>');
@@ -55,6 +62,7 @@
     var applicableOffice = ["18153", "18356", "21453563"];
 
     if (applicableOffice.includes(officeName)) {
+
         jq$(".fullWidthTable").each(function () {
                 //Insert Select Title
                 jq$('<td class="goldTableTitle"> Select </td>').prependTo(jq$(this).find("tr").eq(1));
@@ -89,216 +97,95 @@
     });
 
 
-        /*********Functions**********/
 
-        var pos;
-        //".on" function delegate previous DOM objects to its current modified version (can't use .click)
-        jq$('.content-list').on("click",".flex-list-items", function(){
-            //Move selected user to right of table
-            pos = findPos(names.assignee, jq$(this).text());
 
-            if(jq$(this).hasClass('list-user')){
-                jq$(this).replaceWith("<li class='flex-list-items list-user'></li>");
-                jq$('#assigned-list > li:eq(' + pos + ')').replaceWith(this);
-                jq$(this).addClass('list-assigned').removeClass('list-user');
-            }
+            /*********Functions**********/
 
-            //Move selected user to left of table
-            else if(jq$(this).hasClass('list-assigned')){
-                jq$(this).replaceWith("<li class='flex-list-items list-assigned'></li>");
-                jq$('#user-list > li:eq(' + pos + ')').replaceWith(this);
-                jq$(this).addClass('list-user').removeClass('list-assigned');
-            }
-        });
+            var pos;
+            //".on" function delegate previous DOM objects to its current modified version (can't use .click)
+            jq$('.content-list').on("click",".flex-list-items", function(){
+                //Move selected user to right of table
+                pos = findPos(names.assignee, jq$(this).text());
 
-        //Insert scroll up button
-        jq$('<button id="scrollTop">↑</button>').insertAfter("h1");
-        jq$("#scrollTop").click(function (e) {
-            jq$("html, body").animate({
-                scrollTop: jq$("h1"),
+                if(jq$(this).hasClass('list-user')){
+                    jq$(this).replaceWith("<li class='flex-list-items list-user'></li>");
+                    jq$('#assigned-list > li:eq(' + pos + ')').replaceWith(this);
+                    jq$(this).addClass('list-assigned').removeClass('list-user');
+                }
+
+                //Move selected user to left of table
+                else if(jq$(this).hasClass('list-assigned')){
+                    jq$(this).replaceWith("<li class='flex-list-items list-assigned'></li>");
+                    jq$('#user-list > li:eq(' + pos + ')').replaceWith(this);
+                    jq$(this).addClass('list-user').removeClass('list-assigned');
+                }
             });
-            return false;
-        });
+
+            //Insert scroll up button
+            jq$('<button id="scrollTop">↑</button>').insertAfter("h1");
+            jq$("#scrollTop").click(function (e) {
+                jq$("html, body").animate({
+                    scrollTop: jq$("h1"),
+                });
+                return false;
+            });
 
 
-        //Bulk Button Controls
-        jq$("#moveAllRight").click(function(){
-            jq$(".list-user").remove();
-            jq$(".list-assigned").remove();
-            addNameList("assign");
+            //Bulk Control Buttons
+            // jq$("assigneeList").click(function(){
 
-        });
+            // });
 
-        jq$("#moveAllLeft").click(function(){
-            jq$(".list-user").remove();
-            jq$(".list-assigned").remove();
-            addNameList("default");
-        });
+            // jq$("wfOwnerList").click(function(){
 
-        jq$("#applySubmit").click(function(){
-            var finalized_name = mapFromElement(jq$('.list-assigned'));
-            makePostCall('addTaskAssignee', finalized_name, tableID, selected_office, selected_cu);
-            window.location.reload();
-        });
+            // });
 
+            jq$("#moveAllRight").click(function(){
+                jq$(".list-user").remove();
+                jq$(".list-assigned").remove();
+                addNameList("assign");
 
-        function makePostCall(methodName, assignee, eventID, program_office, costing_unit){
-            if(isNaN(costing_unit)){
-                costing_unit = "";
-            }
+            });
 
-            //Static data
-            var postData = {
-                method: methodName,
-                programOfficeId: program_office,
-                displayPage: false,
-                costingUnitId: costing_unit,
-                roleId: 0,
-            };
+            jq$("#moveAllLeft").click(function(){
+                jq$(".list-user").remove();
+                jq$(".list-assigned").remove();
+                addNameList("default");
+            });
 
-            //Dynamic data
-            var eventIndex;
-            jq$.each(assignee, function(name, id){
-                for (var key = 0; key < eventID.length; key++) {
-                    postData['processIndex'] = key
-                    postData['processOverseer[' + key + ']'] = "";
-                    for(var value = 0; value < eventID[key].length; value++){
-                        eventIndex = eventID[key][value];
-                        postData['taskIndex'] = eventIndex
-                        postData['taskAssigneeId'] = id;
-                        postData['taskAssigneesInfo[' + key + '].selectedRole[' + eventIndex + ']'] = "";
-                        postData['taskAssigneesInfo[' + key + '].selectedAssignee[' + eventIndex + ']'] = id;
-                        postData['taskAssigneesInfo[' + key + '].ccMail[' + eventIndex + ']'] = "";
+            jq$("#applySubmit").click(function(){
+                var finalized_name = mapFromElement(jq$('.list-assigned'));
+                makePostCall('addTaskAssignee', finalized_name, tableID, selected_office, selected_cu);
+                var save_data = {
+                    method: "update",
+                    programOfficeId: selected_office,
+                    displayPage: false,
+                    costingUnitId: selected_cu,
+                    processIndex: 0,
+                    taskIndex: 0,
+                    roleId: 0,
+                    taskAssigneeId: 0,
+                };
 
-                        jq$.ajax({
-                            type: "POST",
-                            async: false,
-                            data: postData,
-                            url: 'https://'
-                            + ((d1$.isEmpty(d1$.hostname)) ? window.location.hostname : d1$.hostname)
-                            +'/srs/sysadmin/system/programOfficeTaskSetup.do?',
+                //Save data after adding users
+                jq$.ajax({
+                    type: "POST",
+                    async: false,
+                    data: save_data,
+                    url: 'https://'
+                    + ((d1$.isEmpty(d1$.hostname)) ? window.location.hostname : d1$.hostname)
+                    +'/srs/sysadmin/system/programOfficeTaskSetup.do?',
 
-                            success : function(data) {
-                                if(debugging) console.log('Data: '+ data);
-                            },
-                            error : function(msg)
-                            {
-                                if(debugging) console.log("Update FAILED: "+ msg);
-                            }
-                        });
+                    success : function(data) {
+                        if(debugging) console.log('Data: '+ data);
+                    },
+                    error : function(msg)
+                    {
+                        if(debugging) console.log("Update FAILED: "+ msg);
                     }
-                }
+                });
+                window.location.reload();
             });
-        };
-
-
-        //default = push all names to left of table
-        //assign = push all names to right of table
-        function addNameList(mode){
-            if(mode == "default"){
-                jq$.each(names.assignee, function(name, id){
-                    jq$('<li>', {
-                    class: 'flex-list-items list-user',
-                    value: id,
-                    text: name,
-                    }).appendTo('#user-list');
-                });
-                jq$.each(names.assignee, function(){
-                    jq$('<li>', {
-                    class: 'flex-list-items list-assigned',
-                    }).appendTo('#assigned-list');
-                });
-            }
-
-            else if(mode == "assign"){
-                    jq$.each(names.assignee, function(){
-                        jq$('<li>', {
-                        class: 'flex-list-items list-user',
-                        }).appendTo('#user-list');
-                    });
-                    jq$.each(names.assignee, function(name, id){
-                        jq$('<li>', {
-                        class: 'flex-list-items list-assigned',
-                        value: id,
-                        text: name,
-                        }).appendTo('#assigned-list');
-                });
-            }
-        }
-
-
-        //Find index of keys from dictionary
-        function findPos(dict, target){
-            var index = 1;
-            for(var name in dict){
-                if(name == target){
-                    break;
-                }
-                ++index;
-            }
-            return index;
-        }
-
-        //Extract names from dropdown box
-        function mapAssignee(program_office, costing_unit){
-            var assignee_selector = null;
-            var wf_owner_selector = null;
-
-            if(program_office == 5){
-                assignee_selector = jq$('[name="taskAssigneesInfo[0].selectedAssignee[2]"] option');
-            }
-
-            //These program office have the same html tag names
-            else if(program_office == 18153 || program_office == 18356){
-                //value of first costing unit if exist is NaN
-                if(isNaN(costing_unit)){
-                    assignee_selector = jq$('[name="taskAssigneesInfo[16].selectedAssignee[1]"] option');
-                    wf_owner_selector = jq$('[name="processOverseer[16]"] option');
-                }
-                else{
-                    assignee_selector = jq$('[name="taskAssigneesInfo[0].selectedAssignee[1]"] option');
-                    wf_owner_selector = jq$('[name="processOverseer[0]"] option');
-                }
-            }
-
-            else if(program_office == 4){
-                assignee_selector = jq$('[name="costingUnitDirectors[0]"] option');
-            }
-
-            else if(program_office == 3){
-                assignee_selector = jq$('[name="taskAssigneesInfo[0].selectedAssignee[1]"] option');
-                wf_owner_selector = jq$('[name="processOverseer[0]"] option');
-            }
-
-            else if(program_office == 2){
-                assignee_selector = jq$('[name="taskAssigneesInfo[1].selectedAssignee[1]"] option');
-                wf_owner_selector = jq$('[name="processOverseer[2]"] option');
-            }
-
-            else if(program_office == 1){
-                assignee_selector = jq$('[name="taskAssigneesInfo[50].selectedAssignee[1]"] option');
-                wf_owner_selector = jq$('[name="processOverseer[50]"] option');
-            }
-
-            var assignee = mapFromElement(assignee_selector);
-            var wf_owner = mapFromElement(wf_owner_selector);
-
-
-            return {"assignee": assignee, "wf_owner" : wf_owner};
-        }
-
-
-        //Convert text and value to a list
-        function mapFromElement(HTML_selector){
-            var result = {};
-
-            jq$(HTML_selector).each(function(){
-                if(jq$(this).text() != ""){
-                    result[jq$(this).text()] = jq$(this).val();
-                }
-            })
-            return result;
-        }
 
 })(jQuery);
