+++ manifest.json
@@ -133,7 +133,8 @@
         "https://*/srs/sysadmin/system/programOfficeTaskSetup.do*"
       ],
       "css": ["d1.workflow.css"],
-      "js": ["d1.workflow.frontend.js"]
+      "js": ["d1.workflow.frontend.js", "d1.workflow.util.js"],
+      "type": "module"
     }
   ],
   "web_accessible_resources": ["d1.common.css","bootstrapiso.min.css"]
